Change BLOCK_SIZE to usize

+++ src/main.rs
@@ -8,7 +8,7 @@ use tokio::fs::File;
 mod bucket;
 mod generate;
 
-const BLOCK_SIZE: u64 = 4096;
+const BLOCK_SIZE: usize = 4096;
 const ONE_GIB: u64 = 1073741824;
 
 /// Generate & sort big files.
@@ -38,7 +38,7 @@ struct Config {
 #[tokio::main]
 async fn main() -> io::Result<()> {
     let cfg = Config::parse();
-    if cfg.max_mem < BLOCK_SIZE {
+    if (cfg.max_mem as usize) < BLOCK_SIZE {
         eprintln!("Max allowed memory must be larger than {BLOCK_SIZE}B");
         process::exit(1);
     }
